fix(agent): tighten admission helper boundaries
